Add files via upload
Agilant App is a Windows Form App (.NET Framework) that was created for the logistics coordinator at Agilant Solutions Inc.
The main purpose for this app is to allow users (mainly technicians) to input in data that using a barcode scanner ex: Asset Tag number, windows version, etc.
After saving the entries the data will be stored in a database. The database created for this app is stored in cloud storage using AWS (Amazon Web Services).
This app is fairly simple to use. (Unfortunately this app did not kickoff due to specific reasons.)

In the ConnectDB directory the file "ConnectDB" is the source code. Caution: There is sensitive information (ex: username and password).